Fork at version 5.0.0 - first thing we do is remove all the modules
One focus of Wily will be to have more flexible modules. The first step
is to strip them out of the main engine's repository, and reintroduce
them separately.

The version 5 release seems like a good time to make a clean break. From
now on we will port only changes to the core engine from upstream.